requirements.txt
pathway
python-dotenv
google-genai
openai
pandas
faker
